% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplotmesh.R
\name{autoplot.inla.mesh}
\alias{autoplot.inla.mesh}
\title{An autoplot (ggplot2) method for INLA mesh objects.}
\usage{
\method{autoplot}{inla.mesh}(mesh, col = "blue", lwd = 0.5, size = 0)
}
\arguments{
\item{mesh}{An inla.mesh object}
}
\description{
An autoplot (ggplot2) method for INLA mesh objects.
}
\examples{
m = 100
points = matrix(runif(m*2),m,2)
mesh = inla.mesh.create.helper(
  points=points,
  cutoff=0.05,
  offset=c(0.1,0.4),
  max.edge=c(0.05,0.5) )

autoplot(mesh)

}

