% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.inla}
\alias{autoplot.inla}
\title{Autoplot method for INLA objects}
\usage{
\method{autoplot}{inla}(x, which = c(1, 2))
}
\arguments{
\item{x}{An inla object}

\item{which}{Vector of integers selecting which plots (1 -- 2) are wanted.}
}
\description{
Replicate the plots produced by \link{\code{plot}} using ggplot.
}
\examples{
 data(Epil)
 my.center = function(x) (x - mean(x))
 ## make centered covariates
 Epil$CTrt    = my.center(Epil$Trt)
 Epil$ClBase4 = my.center(log(Epil$Base/4))
 Epil$CV4     = my.center(Epil$V4)
 Epil$ClAge   = my.center(log(Epil$Age))
 Epil$CBT     = my.center(Epil$Trt*Epil$ClBase4)
 ##Define the model
 formula = y ~ ClBase4 + CTrt + CBT+ ClAge + CV4 +
          f(Ind, model="iid") + f(rand,model="iid")
 result = inla(formula,family="poisson", data = Epil)
 summary(result)
 plot(result)

 p <- autoplot(result)

 p

 p + theme_bw()

 p[2] <- p[2] + ggtitle('Hyper parameters')
 p
}

