% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.inla.residuals.R
\name{ggplot.inla.residuals}
\alias{ggplot.inla.residuals}
\title{Plot residuals of observed vs predicted values for INLA model using ggplot2}
\usage{
\method{ggplot}{inla.residuals}(inla.model, observed, binwidth = NULL)
}
\arguments{
\item{inla.model}{An inla object}

\item{observed}{The observed values}
}
\description{
Plot residuals of observed vs predicted values for INLA model using ggplot2
}
\examples{
 data(Epil)
 observed <- Epil[1:30, 'y']
 Epil <- rbind(Epil, Epil[1:30, ])
 Epil[1:30, 'y'] <- NA
 ## make centered covariates
 formula = y ~ Trt + Age + V4 +
          f(Ind, model="iid") + f(rand,model="iid")
 result = inla(formula, family="poisson", data = Epil, control.predictor = list(compute = TRUE))
 ggplot.inla.residuals(result, observed)
}

