% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.inla.residuals.R
\name{ggplot_inla_residuals2}
\alias{ggplot_inla_residuals2}
\title{Plot residuals against covariate values for INLA model using ggplot2}
\usage{
ggplot_inla_residuals2(inla.model, observed, CI = FALSE)
}
\arguments{
\item{inla.model}{An inla object}

\item{observed}{The observed values}

\item{CI}{plot credible intervals for each residual}
}
\description{
Plot residuals against covariate values for INLA model using ggplot2
}
\examples{
 library(INLA)
 data(Epil)
 observed <- Epil[1:30, 'y']
 Epil <- rbind(Epil, Epil[1:30, ])
 Epil[1:30, 'y'] <- NA
 ## make centered covariates
 formula = y ~ Trt + Age + V4 +
          f(Ind, model="iid") + f(rand,model="iid")
 result = inla(formula, family="poisson", data = Epil, control.predictor = list(compute = TRUE))
 ggplot_inla_residuals2(result, observed)
}

