% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLAutils.R
\docType{package}
\name{INLAutils}
\alias{INLAutils}
\alias{INLAutils-package}
\title{Utility Functions for INLA: Additional Plots and Support for ggplot2.}
\description{
Utility Functions for INLA: Additional Plots and Support for ggplot2.
}
\author{
as.person(c(
   "Tim Lucas <timcdlucas@gmail.com> [aut, cre]", 
   "Dave Redding <dwredding@gmail.com> [aut]"
 ))
}
