% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGAM.R
\name{makeGAM}
\alias{makeGAM}
\title{A function to create the formulae needed to fit GAMs in R.}
\usage{
makeGAM(varnames, y = "y", invariant = "0 + Intercept", linear = TRUE,
  returnstring = TRUE)
}
\arguments{
\item{varnames}{Character vector giving the names of the variables to be used.}

\item{y}{String giving the name of the response variable}

\item{invariant}{Any string to be invluded at the beginning of the formula such as identifying the intercept}

\item{linear}{If TRUE, all variables are included as linear terms. If a character vector, variables in the 
vector are included as linear terms. If FALSE, no linear terms are included.}

\item{returnstring}{If TRUE, return formula as a string (which can later be turned into a formula with 
\code{\link{formula}}). If FALSE, return a formula object.}
}
\description{
Given the beginning of a formula (e.g. y ~ 0 + Intercept) this function builds the string needed to create 
GAM models either with or withour the linear terms as well.
This can then be returned as a formula (if it is to be used directly) or as a string (if other elements such
as random effects are yet to be added).
}

