% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmethods.R
\name{plot.inla.mesh.segment}
\alias{plot.inla.mesh.segment}
\title{These are plot methods for a number of inla objects that do not have plot methods in the INLA package}
\usage{
\method{plot}{inla.mesh.segment}(x, y, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{y}{not used}

\item{...}{other methods passed to plot.data.frame}
}
\description{
These are plot methods for a number of inla objects that do not have plot methods in the INLA package
}
