% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.random.effects}
\alias{plot.fixed.marginals}
\alias{plot.hyper.marginals}
\alias{plot.random.effects}
\title{Individual plot functions for INLA objects}
\usage{
plot.random.effects(x, type = "line")

plot.fixed.marginals(x)

plot.hyper.marginals(x)
}
\arguments{
\item{x}{An inla object}

\item{type}{Which type of plot? 'boxplot' or 'line'}
}
\description{
Replicate the individual plots produced by \link{\code{plot}} using ggplot.
}
\examples{
 data(Epil)
 ##Define the model
 formula = y ~ Trt + Age + V4 +
          f(Ind, model="iid") + f(rand,model="iid")
 result = inla(formula, family="poisson", data = Epil, control.predictor = list(compute = TRUE))

 plot.random.effects(result)
 plot.random.effects(result, type = 'boxplot')
 plot.fixed.marginals(result)
 plot.hyper.marginals(result)
}

