% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printmethods.R
\name{print.inla.data.stack}
\alias{print.inla.data.stack}
\alias{print.inla.mesh}
\alias{print.inla.mesh.projector}
\alias{print.inla.mesh.segment}
\title{These are print methods for a number of inla objects that do not have print methods in the INLA package}
\usage{
\method{print}{inla.data.stack}(x, ...)

\method{print}{inla.mesh}(x, ...)

\method{print}{inla.mesh.projector}(x, ...)

\method{print}{inla.mesh.segment}(x, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
These are print methods for a number of inla objects that do not have print methods in the INLA package
}

