% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloopoint.R
\name{sloopoint}
\alias{sloopoint}
\title{A function to plot the left-out points during the spatial leave-one-out cross-validation in R using INLA.}
\usage{
sloopoint(points = points, test = test, rad = rad)
}
\arguments{
\item{points}{= cbind of longitude and latitude of the observed points (full sample)}

\item{test}{= cbind of longitude and latitude of the test point(s)}

\item{rad}{= Numeric value giving the radius for the spatial buffer around left-out point's location}
}
\description{
This function plots the left-out point(s) of a spatial leave-one-out cross-validation (SLOO-CV) of one or several models running on INLA.
}
\examples{
\dontrun{
 # sloopoint function
 dataframe<-data.frame(long=runif(100, -40.0, 40.0),lat=runif(100, -40.0, 40.0))
 test<-dataframe[1:10,]
 rad = 1
 
 # run the function
 sloopoint(points = cbind(dataframe$long, dataframe$lat), test = cbind(test$long, test$lat), 
 rad = rad)
 }
}
