% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepINLA.R
\name{stepINLA}
\alias{stepINLA}
\title{Perform stepwise variable selection with INLA}
\usage{
stepINLA(fam1 = "gaussian", dataf, spde, in_stack = NULL,
  invariant = "0 + Intercept", direction = c("forwards", "backwards"),
  include = 1:ncol(dataf), y = NULL, y2 = NULL, powerl = 1, inter = 1,
  thresh = 2, Ntrials = NULL, num.threads = 1)
}
\arguments{
\item{fam1}{String defining the likelihood familiy}

\item{dataf}{A SpatialPointsDataFrame}

\item{spde}{An spde model object for the spatial component}

\item{in_stack}{An inla.data.stack object containing all needed data}

\item{invariant}{The part of the formula that should not change (e.g. the intercept and the spatial component.)}

\item{direction}{string 'forwards' for forward variable selection and 'backwards' for backwards variable elimination.}

\item{include}{Vector of integers to determine which columns in dataf should be used.}

\item{y}{String determining the response variable.}

\item{y2}{String determining the name of the test response data.}

\item{powerl}{Integer up to 3 determining which power terms to include.}

\item{inter}{Integer up to 3 determining how many levels of intereactions to include.}

\item{thresh}{Threshold for whether a new model should replace the old model.}

\item{Ntrails}{Not sure.}

\item{num.thread}{How many threads to use for INLA computation.}
}
\description{
Run forwards or backwards stepwise variable selection with INLA regression.
Use hold data to test new models.
Chose the level of interactions and power terms
Force part of the model to remain in the model (e.g. the spatial term and intercept).
}

