% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplotmesh.R
\name{autoplot.inla.mesh}
\alias{autoplot.inla.mesh}
\title{An autoplot (ggplot2) method for INLA mesh objects.}
\usage{
\method{autoplot}{inla.mesh}(object, ..., col = "blue", lwd = 0.5,
  size = 0)
}
\arguments{
\item{object}{An inla.mesh object}

\item{...}{Other arguments passed to specific methods}

\item{col}{Colour for data points}

\item{lwd}{Line width}

\item{size}{size Size of data points}
>>>>>>> 0a5b7afbbd1e107c86c24c35a2848769b1e2ac19
}
\description{
An autoplot (ggplot2) method for INLA mesh objects.
}
\examples{

library(INLA)

m = 100
points = matrix(runif(m*2),m,2)
mesh = inla.mesh.create.helper(
  points=points,
  cutoff=0.05,
  offset=c(0.1,0.4),
  max.edge=c(0.05,0.5) )

autoplot(mesh)

}

