% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.inla}
\alias{autoplot.inla}
\title{Autoplot method for INLA objects}
\usage{
\method{autoplot}{inla}(x, which = c(1:3, 5))
}
\arguments{
\item{x}{An inla object}

\item{which}{Vector of integers selecting which plots (1 -- 4) are wanted.}
}
\description{
Replicate the plots produced by \code{\link{plot}} using ggplot.
}
\details{
The plot types are
\enumerate{
  \item Marginal plot for fixed effects
  \item Marginal plot for hyper parameters
  \item Marginal line plot for random effects
  \item Marginal boxplot plot for random effects (more informative then 3, but gets big with lots of levels
  \item
}
}
\examples{
 library(INLA)
 data(Epil)
 ##Define the model
 formula = y ~ Trt + Age + V4 +
          f(Ind, model="iid") + f(rand,model="iid")
 result = inla(formula, family="poisson", data = Epil, control.predictor = list(compute = TRUE))
 plot(result)

 p <- autoplot(result)

 p

 p + theme_bw()

 p[2] <- p[2] + ggtitle('Hyper parameters')
 p

}

